Add extra configuration to make it work for Win
Because I'm making a lot of assumptions here with the keyboard layout,
it's probably worth making these files separate, so that I can have
different configurations based on machine setup.

This update will be unwanted for UTM setup for sure, but for now, I'm
focused on fixing up the Windows machine setup.