Add waybar toggle
Note that this may need to move to other location, as I would want to
update the waybar at the top as well, but there is no way to respawn
the one at the top if closed from waybar itself. One possibility is to
use swaync, and another is to create separate setup perhaps using
rofi. In either way, this is not the best approach as of now, but is
better to have this control available.